fix : 예약관련 MyBatis -> JPA #62

+++ src/main/java/com/dogdailylog/booking/dao/BookingMapper.java
@@ -20,9 +20,9 @@ public interface BookingMapper {
 
 	public BookingInfo selectBookingByUserIdAndId(
 			@Param("userId") int userId
-			, @Param("id") int id);
+			, @Param("id") Long id);
 
-	public int deleteBookingById(int id);
+	public int deleteBookingById(Long id);
 
 }
